Split the CGNS finite volume reader in a common base class and a class that is specific to the reading of the CGNS grid for the finite volume solver
