+++ Common/include/geometry/meshreader/CCGNSMeshReaderFVM.hpp
@@ -28,64 +28,16 @@
 
 #pragma once
 
-#ifdef HAVE_CGNS
-#include "cgnslib.h"
-#endif
-
-#include "CMeshReaderBase.hpp"
+#include "CCGNSMeshReaderBase.hpp"
 
 /*!
  * \class CCGNSMeshReaderFVM
  * \brief Reads a CGNS zone into linear partitions for the finite volume solver (FVM).
  * \author: T. Economon
  */
-class CCGNSMeshReaderFVM : public CMeshReaderBase {
+class CCGNSMeshReaderFVM : public CCGNSMeshReaderBase {
  private:
 #ifdef HAVE_CGNS
-  int cgnsFileID;         /*!< \brief CGNS file identifier. */
-  const int cgnsBase = 1; /*!< \brief CGNS database index (the CGNS reader currently assumes a single database). */
-  const int cgnsZone = 1; /*!< \brief CGNS zone index (and 1 zone in that database). */
-
-  int nSections; /*!< \brief Total number of sections in the CGNS file. */
-
-  vector<bool> isInterior; /*!< \brief Vector of booleans to store whether each section in the CGNS file is an interior
-                              or boundary section. */
-  vector<unsigned long>
-      nElems; /*!< \brief Vector containing the local number of elements found within each CGNS section. */
-  vector<unsigned long> elemOffset;    /*!< \brief Global ID offset for each interior section (i.e., the total number of
-                                          global elements that came before it). */
-  vector<vector<cgsize_t> > connElems; /*!< \brief Vector containing the local element connectivity found within each
-                                          CGNS section. First index is the section, second contains the connectivity in
-                                          format [globalID VTK n1 n2 n3 n4 n5 n6 n7 n8] for each element. */
-  vector<vector<char> > sectionNames;  /*!< \brief Vector for storing the names of each boundary section (marker). */
-
-  /*!
-   * \brief Open the CGNS file and checks for errors.
-   * \param[in] val_filename - string name of the CGNS file to be read.
-   */
-  void OpenCGNSFile(const string& val_filename);
-
-  /*!
-   * \brief Reads all CGNS database metadata and checks for errors.
-   */
-  void ReadCGNSDatabaseMetadata();
-
-  /*!
-   * \brief Reads all CGNS zone metadata and checks for errors.
-   */
-  void ReadCGNSZoneMetadata();
-
-  /*!
-   * \brief Reads the grid points from a CGNS zone into linear partitions across all ranks.
-   */
-  void ReadCGNSPointCoordinates();
-
-  /*!
-   * \brief Reads the metadata for each CGNS section in a zone and collect information, including the size and whether
-   * it is an interior or boundary section.
-   */
-  void ReadCGNSSectionMetadata();
-
   /*!
    * \brief Reads the interior volume elements from one section of a CGNS zone into linear partitions across all ranks.
    * \param[in] val_section - CGNS section index.
@@ -108,13 +60,6 @@ class CCGNSMeshReaderFVM : public CMeshReaderBase {
    */
   void ReformatCGNSSurfaceConnectivity();
 
-  /*!
-   * \brief Get the VTK type and string name for a CGNS element type.
-   * \param[in] val_elem_type - CGNS element type.
-   * \param[out] val_vtk_type - VTK type identifier index.
-   * \returns String containing the name of the element type.
-   */
-  string GetCGNSElementType(ElementType_t val_elem_type, int& val_vtk_type);
 #endif
 
   /*!
