+++ Common/src/geometry/meshreader/meson.build
@@ -1,4 +1,5 @@
 common_src += files(['CBoxMeshReaderFVM.cpp',
+                     'CCGNSMeshReaderBase.cpp',
                      'CCGNSMeshReaderFVM.cpp',
                      'CMeshReaderBase.cpp',
                      'CRectangularMeshReaderFVM.cpp',
